[addons][audiodecoder] separate "C" to own file
This to hold "C++" and "C" independent and allow for other languages.
