+++ xbmc/addons/kodi-addon-dev-kit/include/kodi/addon-instance/AudioDecoder.h
@@ -10,53 +10,9 @@
 
 #include "../AddonBase.h"
 #include "../AudioEngine.h"
+#include "../c-api/addon-instance/audio_decoder.h"
 
-#include <stdint.h>
-
-extern "C"
-{
-
-typedef struct AddonProps_AudioDecoder
-{
-  int dummy;
-} AddonProps_AudioDecoder;
-
-typedef struct AddonToKodiFuncTable_AudioDecoder
-{
-  KODI_HANDLE kodiInstance;
-} AddonToKodiFuncTable_AudioDecoder;
-
-struct AddonInstance_AudioDecoder;
-typedef struct KodiToAddonFuncTable_AudioDecoder
-{
-  KODI_HANDLE addonInstance;
-  bool(__cdecl* init)(const AddonInstance_AudioDecoder* instance,
-                      const char* file,
-                      unsigned int filecache,
-                      int* channels,
-                      int* samplerate,
-                      int* bitspersample,
-                      int64_t* totaltime,
-                      int* bitrate,
-                      enum AudioEngineDataFormat* format,
-                      const enum AudioEngineChannel** info);
-  int  (__cdecl* read_pcm)(const AddonInstance_AudioDecoder* instance, uint8_t* buffer, int size, int* actualsize);
-  int64_t  (__cdecl* seek)(const AddonInstance_AudioDecoder* instance, int64_t time);
-  bool (__cdecl* read_tag)(const AddonInstance_AudioDecoder* instance,
-                           const char* file, char* title,
-                           char* artist, int* length);
-  int  (__cdecl* track_count)(const AddonInstance_AudioDecoder* instance, const char* file);
-} KodiToAddonFuncTable_AudioDecoder;
-
-typedef struct AddonInstance_AudioDecoder
-{
-  struct AddonProps_AudioDecoder* props;
-  struct AddonToKodiFuncTable_AudioDecoder* toKodi;
-  struct KodiToAddonFuncTable_AudioDecoder* toAddon;
-} AddonInstance_AudioDecoder;
-
-} /* extern "C" */
-
+#ifdef __cplusplus
 namespace kodi
 {
 namespace addon
@@ -366,3 +322,4 @@ private:
 
 } /* namespace addon */
 } /* namespace kodi */
+#endif /* __cplusplus */
